DDPB-2355 moved estimates properties into dedicated trait
